requirements2.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
dataclasses==0.8
Django==3.2.7
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.2
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
pyttsx3==2.90
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.2
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
